NexoraCode/requirements.txt
pywebview>=5.0
flask>=3.0
flask-cors>=4.0
requests>=2.31
pystray>=0.19
Pillow>=10.0
playwright>=1.43
trafilatura>=1.9
